+++ Source/Core/Platform/Platform.h
@@ -59,7 +59,7 @@ RIFT_API void* Platform_MemSet(void* Dest, i32 Value, usize Size);
 RIFT_API bool Platform_MemEqual(const void* Block1, const void* Block2, usize Size);
 
 RIFT_API void Platform_ConsoleWrite(const char* Message, u8 Color, bool bIsError);
-RIFT_API void Platform_ConsoleWrite_CustomLength(const char* Message, u64 Length, u8 Color, bool bIsError);
+RIFT_API void Platform_ConsoleWrite_CustomLength(const char* Message, u32 Length, u8 Color, bool bIsError);
 
 RIFT_API PlatformHandle Platform_CreateThread(const String Name, u32* OutThreadID, u32 (*ThreadEntryPoint)(void* ThreadParameter), void* UserData);
 RIFT_API PlatformHandle Platform_RunCommand(const String CmdLine, const String WorkingDirectory);
